Fix for incorrect partition detection logic Contributes to JB#48714
Original code of find-mmc-bypartlabel looks for any mmc partition
device appearance in sysfs and waits if there is none. Otherwise
it directly checks for label it is looking for and if not found
exits with error.

However on mmcs with large number of partitions (or slower ones)
partitions do not appear all at once, so it is possible that
1. There are mmc partitions in sysfs
2. There is no particular label partition yet
(because not all partitions are there yet).

The patch here handles this partcular case in pecular way
by waiting a bit and retrying 50 times (which makes worse case
till error to 5 seconds which should be enough for most cases,
but this might need to be increased in case of marginally slow
EMMC chips (or old heavily used EMMC chips).

Contributes to JB#48714