[factory-reset] Differentiate error cases. JB#63686
Return 1 on "soft" failures, i.e. when the factory image is not found or
does not pass verification, and no modifications have been made to the
system yet. And return 2 if the reset fails at point where the system is
most likely in a broken state.

This allows clearing the flag file and booting back to the old system on
soft failures.